third_party/WebKit/LayoutTests/nfc/mock-nfc.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/nfc-helpers.js"></script>
<script>

'use strict';

nfc_test(nfc => {
  assert_true(nfc instanceof Object);
  assert_true(nfc.mockNFC instanceof Object);
}, 'NFC Mojo bindings and mock interfaces are available to tests.')
</script>
